requirements.txt
pillow
cvzone
opencv-python
numpy
python-bidi
arabic-reshaper